Fixes shop collection page
Shop collection page was crashing on reloading. The reason was the collections state value, which was null. So before becoming isFetching, App tried to .map null value of collections.

This commit adds a selector which converts collections to Boolean.
Now, with this prop being passed to WithSpinner HOC, App doesn't crash on page reload.